requirements.txt
APScheduler==3.6.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
binance-connector==2.0.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==40.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
service-identity==21.1.0
six==1.16.0
tornado==6.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
zope.interface==6.0
